Move Expr into GL.Shader.DSL.
